backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
redis==5.2.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
openai==1.57.4
PyJWT[crypto]==2.10.1
cryptography==44.0.0
python-multipart==0.0.17
PyYAML==6.0.2
